requirements.txt
google-generativeai>=0.3.0
requests>=2.28.1
python-dotenv>=1.0.0
beautifulsoup4>=4.11.1
markdown>=3.4.1
selenium>=4.5.0
webdriver-manager>=3.8.5
unidecode>=1.3.6